Rewritten docstrings, added expected parameters type, added explicit function return value, added getters/setters instead of using private variables. Using Google YAPF to format python files

+++ ava/plugins/plugin.py
@@ -1,9 +1,10 @@
-from os import path
-from .process import spawn
+from os.path import join
 from subprocess import Popen
-# from ..state import State
-# from ..loading import loading
 from avasdk.plugins.utils import load_plugin
+from .process import spawn
+from ..state import State
+
+__all__ = ['State']
 
 
 class Plugin(object):
@@ -16,8 +17,8 @@ class Plugin(object):
         We load the plugin's manifest and spawn a process in which the execution
         of the plugin will be contained.
 
-        :name: Plugin name (string).
-        :path: Path towards the plugin folder containing the code (string).
+        :param name: Plugin name.
+        :param path: Path towards the plugin folder containing the code.
         """
         self._name = name
         self._path = path
@@ -25,29 +26,30 @@ class Plugin(object):
         self._process = spawn(self)
 
     def __repr__(self):
-        return f'<AVA.plugin.name:{self._name.capitalize()}>'
+        return f'<{self.__class__.__module__}.{self.__class__.__name__} {self._name.capitalize()} at {hex(id(self))}>'
 
     def get_name(self) -> str:
         """
-        :return: A string containing the name of the plugin.
+        :return: Returns a string containing the name of the plugin.
         """
         return self._name
 
     def get_path(self) -> str:
         """
-        :retrurn: Path towards the plugin folder containing the source code.
+        :retrurn: Returns the path towards the plugin folder containing the
+         source code.
         """
-        return path.join(self._path, self._name)
+        return join(self._path, self._name)
 
     def get_process(self) -> Popen:
         """
-        :return: The subprocess.Popen instance
+        :return: Returns the subprocess.Popen instance.
         """
         return self._process
 
     def get_specs(self) -> dict:
         """
-        :return: The plugin's manifest.json as dictionary.
+        :return: Returns the plugin's manifest.json as a dictionary.
         """
         return self._specs
 
@@ -59,8 +61,8 @@ class Plugin(object):
         the process and set back 'self._process' to None.
         It ensures to properly restart the plugin after that.
         """
-        assert self._process is not None
-        self._process.kill()
+        if self._process is not None and isinstance(self._process, Popen):
+            self._process.kill()
         self._process = None
 
     def restart(self):
@@ -70,20 +72,17 @@ class Plugin(object):
         We spawn a new process dedicated to this plugin and we store it into
         'self._process' to make it available and usable again.
         """
-        assert self._process is None
-        # State().loading()
-        # loading(plugins_nbr=1, process_time=6, target=self.name)
-        self._process = spawn(self)
-        # State().loading_done()
+        if self._process is None:
+            self._process = spawn(self)
 
     def shutdown(self):
         """
-        Shutdown the plugin gracefully.
+        Shutdown gracefully the plugin.
 
         In order to properly shutdown a plugin, we call the 'terminate()'
         method of the subprocess.Popen object and set back 'self.process' to
         None.
         """
-        assert self._process is not None
-        self._process.terminate()
+        if self._process is not None and isinstance(self._process, Popen):
+            self._process.terminate()
         self._process = None
